Add authorization scopes
