+++ zapier/index.js
@@ -1,4 +1,4 @@
-const authentication = require('./authentication');
+const auth = require('./auth');
 const executeWorkflow = require('./creates/executeWorkflow');
 const hydrators = require('./hydrators');
 const listModels = require('./triggers/listModels');
@@ -6,30 +6,12 @@ const listTrainings = require('./triggers/listTrainings');
 const listWorkflows = require('./triggers/listWorkflows');
 const postPrediction = require('./creates/postPrediction');
 
-const addAuthorization = async (request, z, bundle) => {
-  const basicHash = Buffer.from(
-    `${bundle.authData.app_client_id}:${bundle.authData.app_client_secret}`
-  ).toString(
-    'base64'
-  );
-  const response = await fetch(process.env.API_AUTH_URL, {
-    method: 'POST',
-    headers: {
-      'Authorization': `Basic ${basicHash}`,
-      'Content-Type': 'application/x-www-form-urlencoded',
-    },
-  });
-  data = await response.json()
-  request.headers.Authorization = `Bearer ${data.access_token}`;
-  return request;
-};
-
 module.exports = {
   version: require('./package.json').version,
   platformVersion: require('zapier-platform-core').version,
-
-  authentication,
-  beforeRequest: [addAuthorization],
+  
+  authentication: auth.authentication,
+  beforeRequest: [auth.addAuthorization],
   hydrators,
 
   triggers: {
